--- a/tests/cupy_tests/statics_tests/test_correlation.py
+++ b/tests/cupy_tests/statics_tests/test_correlation.py
@@ -111,6 +111,8 @@ class TestCorrelateShapeCombination(unittest.TestCase):
 @testing.parameterize(*testing.product({
     'mode': ['valid', 'full', 'same']
 }))
+@unittest.skipIf(
+    cupy.cuda.cufft.getVersion() == 10000, 'Causes large precision errors')
 class TestCorrelate(unittest.TestCase):
 
     @testing.for_all_dtypes()
@@ -121,18 +123,17 @@ class TestCorrelate(unittest.TestCase):
         return xp.correlate(a[::200], b[10::70], mode=self.mode)
 
     @testing.for_all_dtypes(no_float16=True)
-    @testing.numpy_cupy_allclose(rtol=1e-3)
+    @testing.numpy_cupy_allclose(rtol=1e-4)
     def test_correlate_large_non_contiguous(self, xp, dtype):
         a = testing.shaped_arange((10000,), xp, dtype)
         b = testing.shaped_arange((1000,), xp, dtype)
         return xp.correlate(a[200::], b[10::700], mode=self.mode)
 
-    @testing.for_all_dtypes_combination(
-        names=['dtype1', 'dtype2'])
+    @testing.for_all_dtypes_combination(names=['dtype1', 'dtype2'])
     @testing.numpy_cupy_allclose(rtol=1e-2)
     def test_correlate_diff_types(self, xp, dtype1, dtype2):
-        a = testing.shaped_arange((200,), xp, dtype1)
-        b = testing.shaped_arange((100,), xp, dtype2)
+        a = testing.shaped_random((200,), xp, dtype1)
+        b = testing.shaped_random((100,), xp, dtype2)
         return xp.correlate(a, b, mode=self.mode)