Rel. #3556.
Blocked by #3571.

Previous:
```
convolve            :    CPU: 1598.203 us   +/-196.056 (min: 1553.705 / max: 5673.330) us     GPU-0: 1607.828 us   +/-198.819 (min: 1562.784 / max: 5725.760) us
convolve            :    CPU: 7661.681 us   +/-683.048 (min: 7492.259 / max:14718.648) us     GPU-0: 7673.338 us   +/-684.819 (min: 7503.136 / max:14733.984) us
convolve            :    CPU: 7556.341 us   +/-45.591 (min: 7495.102 / max: 7804.286) us     GPU-0: 7567.647 us   +/-45.571 (min: 7506.176 / max: 7815.456) us
convolve            :    CPU: 9845.364 us   +/-53.787 (min: 9777.471 / max:10081.744) us     GPU-0: 9856.696 us   +/-53.819 (min: 9788.608 / max:10093.184) us
```
This PR (https://github.com/cupy/cupy/pull/3560/commits/b570a128a6bf1a76136fedc5ba2b7abcc763f43e):
```
convolve            :    CPU: 1428.156 us   +/-157.454 (min: 1391.433 / max: 5490.528) us     GPU-0: 1435.945 us   +/-158.645 (min: 1399.072 / max: 5529.664) us
convolve            :    CPU: 1511.592 us   +/-144.757 (min: 1472.319 / max: 4345.249) us     GPU-0: 1519.854 us   +/-147.089 (min: 1480.000 / max: 4380.480) us
convolve            :    CPU: 1505.722 us   +/-32.617 (min: 1474.274 / max: 1744.067) us     GPU-0: 1513.800 us   +/-32.601 (min: 1482.432 / max: 1752.128) us
convolve            :    CPU: 1509.725 us   +/-72.535 (min: 1474.584 / max: 3164.543) us     GPU-0: 1518.026 us   +/-75.407 (min: 1482.656 / max: 3254.400) us
```

benchmark script:
```py
import numpy

import cupy
import cupyx

for a_len in [10, 200000]:
    for v_len in [10, 200000]:
        a = cupy.random.rand(a_len)
        v = cupy.random.rand(v_len)
        perf = cupyx.time.repeat(
            cupy.convolve, (a, v, 'same'),
            n_warmup=5, max_duration=1)
        print(perf)
```