@emcastillo 2020-07-07
The PR pads the size to be the closest power of 2 so the fft algorithms can work more efficiently.