@asi1024 2020-07-07
`cupy-py35` which uses cuFFT 10.1.2 has passed, and `cupy-py3` which uses cuFFT 10.0.0 has failed.