@grlee77 2020-07-07
@emcastillo 
Good idea to update the size to a favorable one.

Note, that you don't necessarily need to increase to a power of two, but could instead go to the next smallest number that has only 2, 3, 5, and 7 as its prime factors (ref: https://docs.nvidia.com/cuda/cufft/index.html#accuracy-and-performance). 

SciPy has a utility function for this ([scipy.fft.next_fast_len](https://docs.scipy.org/doc/scipy/reference/generated/scipy.fft.next_fast_len.html)), which has a docstring example where the next power of two makes the signal unnecessarily large. pocketfft's prime factors are slightly different from cuFFT's factors: they are {2, 3, 5, 7, 11} for complex-to-complex transforms, but {2, 3, 5} for transforms involving real input or output. The latter would work fine for cuFFT (just neglecting some potential cases involving factors of 7)

Specifically, if scipy >= 1.4 is available, one can call

```Python
fast_size = scipy.fft.next_fast_len(size, real=True)
```
For older SciPy, FFTPACK only supported factors (2, 3, 5) so
```Python
fast_size = scipy.fftpack.next_fast_len(size)
```
would work instead.

If you really wanted to make a cuFFT-dedicated C++ function also supporting factor 7 you could start from the C++ ones in pocketfft as here:
https://github.com/scipy/scipy/blob/3f11e32f3d07e6035b46f77eafec4359875a0416/scipy/fft/_pocketfft/pocketfft_hdronly.h#L411-L436