@leofang 2020-07-07
> If you really wanted to make a cuFFT-dedicated C++ function also supporting factor 7 you could start from the C++ ones in pocketfft 

I once thought about adding `cupyx.scipy.fft*.next_fast_len()`, but I didn't have strong motivation (until now seeing this). But, due to the difference between pocketfft and cuFFT, the returned values won't be identical. Not sure if this is acceptable.