@grlee77 2020-07-07
> But, due to the difference between pocketfft and cuFFT, the returned values won't be identical. Not sure if this is acceptable.

I would personally find it acceptable, as it is conceptually the same: "find me the next smallest length that will not have poor performance in the underlying FFT library". Particularly in the context of a round-trip transform as for FFT-based convolution, I don't think the user will care what intermediate size was used internally.