@leofang 2020-07-08
Hi @grlee77 You know I almost always agree with you on performance considerations 😁 I was refreshing my memory on SciPy's implementation: https://github.com/scipy/scipy/blob/3f11e32f3d07e6035b46f77eafec4359875a0416/scipy/fft/_helper.py#L6
and your suggestion

> (we can always add a note: "Due to differences in the underlying FFT libraries used by CuPy and SciPy, the length returned will not always be identical to scipy.fft.next_fast_len.")

brings up a question: with the recent backend support in `scipy.fft`, perhaps `scipy.fft.next_fast_len()` should also be registered with a backend? Ideally, we want this usage
```python
import scipy.fft as spfft
import cupyx.scipy.fft as cufft

spfft.set_global_backend(cufft)
n = spfft.next_fast_len(100)
```
to correctly return the number for cuFFT, not for pocketfft.