@leofang 2020-07-08
@asi1024 @grlee77 I don't think I fully understand this, but here you go:
```python 
cpdef size_t next_fast_len(size_t n):
    """2^a * 3^b * 5^c * 7^d for cuFFT."""

    if n <= 8:
        return n

    cdef size_t b = 1, c = 1, d = 1, best = 2 * n

    while d < best:
        c = d
        while c < best:
            b = c
            while b < n:
                b *= 2
            while True:
                if b < n:
                    b *= 3
                elif b > n:
                    if b < best:
                        best = b
                    if b & 1:
                        break
                    b >>= 1
                else:
                    return n
            c *= 5
        d *= 7
    return best
```