@grlee77 2020-07-08
> @asi1024 @grlee77 How about this? This is basically what SciPy does, tailored to primes (2, 3, 5, 7).

Yes, this looks correct. I empirically verified via comparison to `scipy.fft.next_fast_len` by checking that when there are factors of 11 the scipy result is smaller than this one, otherwise they are the same. details below (with `prime_factors` function taken from this [stackoverflow post](https://stackoverflow.com/questions/15347174/python-finding-prime-factors))

<details>

```Python
def prime_factors(n):
    i = 2
    factors = []
    while i * i <= n:
        if n % i:
            i += 1
        else:
            n //= i
            factors.append(i)
    if n > 1:
        factors.append(n)
    return factors


def next_fast_len(n):
    """2^a * 3^b * 5^c * 7^d for cuFFT."""
    if n <= 8:
        return n
    b = 1
    c = 1
    d = 1
    best = 2 * n
    while d < best:
        c = d
        while c < best:
            b = c
            while b < n:
                b *= 2
            while True:
                if b < n:
                    b *= 3
                elif b > n:
                    if b < best:
                        best = b
                    if b & 1:
                        break
                    b >>= 1
                else:
                    return n
            c *= 5
        d *= 7
    return best


for n in range(1024):
    n_fast_scipy = scipy.fft.next_fast_len(a)
    n_fast = next_fast_len(a)
    if 11 in prime_factors(n_fast_scipy):
        assert n_fast_scipy < n_fast
    else:
        assert n_fast_scipy == n_fast
```

</details>