@leofang 2020-07-08
Awesome! Thanks for confirming, @grlee77. 

@asi1024, would you like to include `next_fast_len()` in this PR, or should I send a new one? Note that it better be implemented in C or Cython, because it'd be way faster:
```python
In [3]: %timeit nfl_cython(100000000000)                                                                                                                             
489 ns ± 6.36 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [4]: %timeit nfl_python(100000000000)                                                                                                                               
53.4 µs ± 787 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

In [6]: %timeit nfl_cython(1000000)                                                                                                                                    
176 ns ± 0.897 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [7]: %timeit nfl_python(1000000)                                                                                                                                    
16.2 µs ± 152 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```