@leofang 2020-07-08
@asi1024 Could you rerun your benchmark and see it's better? 😇