@asi1024 2020-07-08
Should we fix to 1. skip tests if `cupy.cuda.cufft.getVersion() == 10000` or 2. relax tolerance to pass with any version of cuFFT?