@leofang 2020-07-08
I am fine either way, was just wondering why we didn't see this issue in individual fft tests. 

@anaruse Do you happen to know if there's known precision issues for 1D FFT in CUDA 10.0?