@asi1024 2020-07-08
Updated performance numbers in the PR description!