@emcastillo 2020-07-09
I measured the dot convolution and the fft depending on the inputs a and b sizes. What I got is this

![image](https://user-images.githubusercontent.com/15824382/87007582-76aa9980-c1fd-11ea-81ea-18c4718c3b32.png)

Y axis is a size, X axis is b size.
In green you can see the results when direct convolution is faster than fft.
This is measured in a V100 GPU so the threshold is likely to change depending on the GPU.