@leofang 2020-07-09
@asi1024 @grlee77 I am wondering for `C2C`  if we could further speed up this by reusing a `Plan1d` for the three fft calls